// Angular/src/app/Models/cars.ts
import { Image } from "./image";


export class Cars {
  id?: number;
  name?: string;
  model?: string;
  nb_doors?: number;
  nb_places?: number;
  address?: string;
  description?: string;
  price_per_day?: number;
  photo?: File;
  image?: Image;
  registration_num?: string;
  gearbox?: string;
}